+++ addon/routes/fd-diagrams.js
@@ -137,8 +137,8 @@ export default Route.extend({
 
       @method actions.willTransition
     */
-    willTransition() {
-      this.get('fdSheetService').closeSheet(this.get('sheetComponentName'));
+    willTransition(transition) {
+      this.get('fdSheetService').transitionFromSheet(transition, this.get('sheetComponentName'));
 
       this._super(...arguments);
     }