+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-### Added
-* Check unsaved data on sheet.
+
+## [0.9.1] - 2019-10-08
+### Fixed
+* Fix sidebar and list styles.
 
 ## [0.9.0] - 2019-10-02
 ### Added